test: split test files and increase test coverage
